+++ lib/index.js
@@ -1,6 +1,10 @@
 /*
- * Copyright (c) 2017-2018 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2018-2019 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
-module.exports = require('./ld-key-pair');
+module.exports = {
+  Ed25519KeyPair: require('./Ed25519KeyPair'),
+  LDKeyPair: require('./LDKeyPair'),
+  RSAKeyPair: require('./RSAKeyPair'),
+};
